+++ packages/react-native/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [13.0.5](https://github.com/mbret/prose-reader/compare/v13.0.4...v13.0.5) (2026-09-25)
+
+
+### Bug Fixes
+
+* **core:** open at the initial cfi without reporting the start of the book ([fd1ea00](https://github.com/mbret/prose-reader/commit/fd1ea00a6d7e8384d1f73a0090ec8c1f26ac1ef9))
+
+
+
+## [13.0.4](https://github.com/mbret/prose-reader/compare/v13.0.3...v13.0.4) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-native
+
+
+
+
+
+## [13.0.2](https://github.com/mbret/prose-reader/compare/v13.0.1...v13.0.2) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-native
+
+
+
+
+
+## [13.0.1](https://github.com/mbret/prose-reader/compare/v13.0.0...v13.0.1) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-native
+
+
+
+
+
+## [13.0.0](https://github.com/mbret/prose-reader/compare/v12.0.0...v13.0.0) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-native
+
+
+
+
+
 ## [12.0.0](https://github.com/mbret/prose-reader/compare/v11.0.0...v12.0.0) (2026-09-24)
 
 **Note:** Version bump only for package @prose-reader/react-native